from-up-IEEE80211-es/requirements.txt
lxml==4.2.3
elasticsearch==7.13.1
pyyaml
requests
requests-aws4auth==0.9
beautifulsoup4==4.9.3
boto3==1.17.88
